CloudWatch Events --> Amazon EventBridge
Update instructions to clarify this lambda supports Amazon EventBridge (formerly known as CloudWatch Events).

Update the code so that it doesn't depend on deprecated dependencies (signalfx-lambda-nodejs-wrapper layer.

Remove examples as with the new approach it is easy to inspect and modify the lambda code as all key parts are now included in this repo and not hidden in the layer code as before.